containers/Deenish-Island/site/requirements.txt
python-dateutil
netCDF4==1.5.7
numpy
paramiko
pytz
pysftp
pandas
plotly
